search: move DetermineStatusForLogs out of graphqlbackend
This was a weird dependency on graphqlbackend in the streaming endpoint.
I chose client as the package for this to live in with the intention to
move and unify search telemetry into the client. So this is a work in
progress and will evolve quickly.

Test Plan: CI
