+++ cmd/frontend/internal/search/search.go
@@ -18,7 +18,6 @@ import (
 	"github.com/sourcegraph/log"
 	"go.opentelemetry.io/otel/attribute"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	searchlogs "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/search/logs"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
@@ -180,7 +179,7 @@ func (h *streamHandler) serveHTTP(r *http.Request, tr *trace.Trace, eventWriter 
 
 func logSearch(ctx context.Context, logger log.Logger, alert *search.Alert, err error, duration time.Duration, latency *time.Duration, originalQuery string, progress *streamclient.ProgressAggregator) {
 	if honey.Enabled() {
-		status := graphqlbackend.DetermineStatusForLogs(alert, progress.Stats, err)
+		status := client.DetermineStatusForLogs(alert, progress.Stats, err)
 		var alertType string
 		if alert != nil {
 			alertType = alert.PrometheusType
